+++ GameServer/Model/Account/AccountEntitie.cs
@@ -2,18 +2,18 @@
 
 namespace GameServer.Model.Account;
 public class AccountEntitie {
-    public int Id { get; set; }
+    public uint Id { get; set; }
     public uint ConnectionId { get; set; }
     public string Username { get; set; }
     public string PasswordHash { get; set; }
     public string Token { get; set; }
     public DateTime TokenCreationTime { get; set; }
-    public int AccountStatus { get; set; }
-    public int BanDays { get; set; }
-    public int Nation { get; set; }
+    public ushort AccountStatus { get; set; }
+    public ushort BanDays { get; set; }
+    public ushort Nation { get; set; }
     public AccountType AccountType { get; set; }
-    public int StorageGold { get; set; }
-    public int Cash { get; set; }
+    public uint StorageGold { get; set; }
+    public uint Cash { get; set; }
     public string PremiumExpiration { get; set; }
     public List<CharacterEntitie> Characters { get; set; } = [];
 }
